first pass, rough decomposition of test helpers file
