+++ packages/datastore/__tests__/helpers/fakes/index.ts
@@ -1,0 +1,2 @@
+export { FakeGraphQLService } from './graphqlService';
+export { FakeDataStoreConnectivity } from './connectivityMonitor';
