+++ packages/datastore/__tests__/helpers/schemas/index.ts
@@ -1,0 +1,4 @@
+export * from './default';
+export * from './internal';
+export * from './small';
+export * from './typeOnlyModels';
